requirements.txt
django==2.1.2
django-heroku==0.3.1
gunicorn==19.9.0
dj_database_url
psycopg2
djangorestframework==3.9.0
channels==2.1.5
channels-redis==2.3.1
Pillow
uvicorn